Refactor tests to utilize get_entity_id_by_unique_id_suffix function from conftest.py
