+++ tests/test_switch.py
@@ -13,6 +13,8 @@ from homeassistant.const import (
 )
 from homeassistant.helpers import entity_registry as er
 
+from .conftest import get_entity_id_by_unique_id_suffix
+
 if TYPE_CHECKING:
     from unittest.mock import MagicMock
 
@@ -25,16 +27,6 @@ SWITCH_DOMAIN = "switch"
 def platforms() -> list[Platform]:
     """Return the platforms to be loaded for this test."""
     return [Platform.SWITCH]
-
-
-def get_entity_id_by_unique_id_suffix(
-    entity_registry: er.EntityRegistry, suffix: str
-) -> str | None:
-    """Get entity_id from the registry by unique_id suffix."""
-    for entry in entity_registry.entities.values():
-        if entry.unique_id and entry.unique_id.endswith(suffix):
-            return entry.entity_id
-    return None
 
 
 @pytest.mark.usefixtures("init_integration")
